Shorten #console in Console
