+++ bootstrap.rb
@@ -1,7 +1,10 @@
+require_relative 'lib/modules_console/money_operations_console'
 require_relative 'lib/console'
+
 require_relative 'lib/modules/bank_errors'
 require_relative 'lib/modules/file_loader'
 require_relative 'lib/modules/money_operations'
+
 require_relative 'lib/cards/basic_card'
 require_relative 'lib/cards/usual_card'
 require_relative 'lib/cards/capitalist_card'
